requirements.txt
fastapi
uvicorn
groq
httpx
aiosqlite
line-bot-sdk>=3.0.0
tzdata
chromadb
pdfplumber
google-genai
python-docx
cachetools

openai
simpleeval

# Dev / test
pytest
pytest-asyncio
